refactor: Move code to specific page
